+++ tests/domain/cases/checkoutPagamento.test.ts
@@ -1,15 +1,15 @@
 import { describe } from 'node:test';
 import { expect } from '@jest/globals';
 import { test, it, beforeAll, afterAll, jest } from '@jest/globals';
-import CheckoutPagamento from '../../../app/domain/cases/checkoutPagamento';
-import CheckoutPagamentoRepository from '../../../app/application/repositories/CheckoutPagamentoRepository';
+import CheckoutPagamento from '../../../app/cases/checkoutPagamento';
+import CheckoutPagamentoRepository from '../../../app/gateways/CheckoutPagamentoRepository';
 import MysqlDataBase from '../../../app/application/database/MysqlDataBase';
-import MPagamento from '../../../app/application/core/paymentsMethods/MercadoPago/MPagamento';
-import PedidoRepository from '../../../app/application/repositories/PedidoRepository';
+import MPagamento from '../../../app/gateways/paymentsMethods/MercadoPago/MPagamento';
+import PedidoRepository from '../../../app/gateways/PedidoRepository';
 import Checkout from '../../../app/domain/entity/checkout';
 import Cartao from '../../../app/domain/entity/cartao';
 import Payer from '../../../app/domain/entity/payer';
-import PaymentoMethods from '../../../app/application/core/paymentsMethods/PaymentoMethods';
+import PaymentoMethods from '../../../app/gateways/paymentsMethods/PaymentoMethods';
 
 
 describe("TEST use case Checkout Pagamento.", () =>{
