+++ tests/domain/cases/usuarioCasoDeUso.test.ts
@@ -1,7 +1,7 @@
 import { describe } from 'node:test';
 import { expect } from '@jest/globals';
 import { test, it, beforeAll, afterAll, jest } from '@jest/globals';
-import UsuarioCasoDeUso from '../../../app/domain/cases/usuarioCasoDeUso';
+import UsuarioCasoDeUso from '../../../app/cases/usuarioCasoDeUso';
 import User from '../../../app/domain/entity/user';
